Split commands/ into one submodule per domain
Decompose the monolithic commands/mod.rs into domain submodules — github,
admin, devcontainer, profiles, quickstart, lifecycle — keeping mod.rs as the
shared orchestration layer (merge_runtime_guest_env, purge_all_data) plus the
pub(crate) re-export block. The dispatch surface lib.rs imports is byte-for-byte
unchanged. Helpers used only across submodules narrow to pub(super); the seven
helpers lib.rs::run calls directly stay pub(crate). Tests move next to the code
they exercise. No behavior change.
